widget:     Introduced the widget object, which contains all gui     rendering instructions.
wlmenu:
    Removed all rendering operations and moved them to the widget interface.

main:
    Changed due to the introduction of widget.

load:
    "items.hits" is now type "size_t" to reduce type casting during the
    item selection.

+++ src/load.h
@@ -25,9 +25,11 @@
 #ifndef LOAD_H_
 #define LOAD_H_
 
+#include <stdint.h>
+
 struct item {
     char *name;
-    unsigned int hits;
+    size_t hits;
 };
 
 size_t load(struct item **list);
